* Move hudson class usage to impl * Introduce JUnit specific subclass of BlueTestResult
