+++ blueocean-rest/pom.xml
@@ -19,10 +19,6 @@
       <groupId>${project.groupId}</groupId>
       <artifactId>blueocean-web</artifactId>
     </dependency>
-    <dependency>
-      <groupId>org.jenkins-ci.plugins</groupId>
-      <artifactId>junit</artifactId>
-    </dependency>
 
       <dependency>
           <groupId>${project.groupId}</groupId>
